feat(ui): move connect wallet screen to app
