+++ src/ui/components/WalletConnect/WalletConnect.scss
