+++ src/ui/components/WalletConnect/WalletConnect.types.ts
@@ -1,4 +1,4 @@
-import { ConnectionData } from "../../../../../store/reducers/walletConnectionsCache";
+import { ConnectionData } from "../../../store/reducers/walletConnectionsCache";
 
 interface WalletConnectStageOneProps {
   isOpen: boolean;
